requirements.txt
websockets==12.0
pygame==2.5.2
pytest==7.4.3
black==23.12.1
ruff==0.1.8
mypy==1.7.1
python-dotenv==1.0.0
